大批小问题修复

+++ drivers/media/platform/msm/camera/cam_isp/isp_hw_mgr/cam_ife_hw_mgr.c
@@ -4013,8 +4013,8 @@ static int cam_isp_packet_generic_blob_handler(void *user_data,
 
 		if (blob_size < (sizeof(uint32_t) * 2 + hfr_config->num_ports *
 			sizeof(struct cam_isp_port_hfr_config))) {
-			CAM_ERR(CAM_ISP, "Invalid blob size %u expected %u",
-				blob_size, sizeof(uint32_t) * 2 +
+			CAM_ERR(CAM_ISP, "Invalid blob size %zu expected %zu",
+				(size_t)blob_size, sizeof(uint32_t) * 2 +
 				sizeof(struct cam_isp_port_hfr_config) *
 				hfr_config->num_ports);
 			return -EINVAL;
@@ -4044,8 +4044,8 @@ static int cam_isp_packet_generic_blob_handler(void *user_data,
 
 		if (blob_size < (sizeof(uint32_t) * 2 + sizeof(uint64_t) *
 			(clock_config->num_rdi + 2))) {
-			CAM_ERR(CAM_ISP, "Invalid blob size %u expected %u",
-				blob_size,
+			CAM_ERR(CAM_ISP, "Invalid blob size %zu expected %zu",
+				(size_t)blob_size,
 				sizeof(uint32_t) * 2 + sizeof(uint64_t) *
 				(clock_config->num_rdi + 2));
 			return -EINVAL;
@@ -4076,8 +4076,8 @@ static int cam_isp_packet_generic_blob_handler(void *user_data,
 
 		if (blob_size < (sizeof(uint32_t) * 2 + (bw_config->num_rdi + 2)
 			* sizeof(struct cam_isp_bw_vote))) {
-			CAM_ERR(CAM_ISP, "Invalid blob size %u expected %u",
-				blob_size,
+			CAM_ERR(CAM_ISP, "Invalid blob size %zu expected %zu",
+				(size_t)blob_size,
 				sizeof(uint32_t) * 2 + (bw_config->num_rdi + 2)
 				* sizeof(struct cam_isp_bw_vote));
 			return -EINVAL;
@@ -4121,8 +4121,8 @@ static int cam_isp_packet_generic_blob_handler(void *user_data,
 		if (blob_size < (sizeof(uint32_t) * 2
 			+ (bw_config_ab->num_rdi + 2)
 			* sizeof(struct cam_isp_bw_vote))) {
-			CAM_ERR(CAM_ISP, "Invalid blob size %u expected %u",
-				blob_size,
+			CAM_ERR(CAM_ISP, "Invalid blob size %zu expected %zu",
+				(size_t)blob_size,
 				sizeof(uint32_t) * 2
 				+ (bw_config_ab->num_rdi + 2)
 				* sizeof(struct cam_isp_bw_vote));
@@ -4164,8 +4164,8 @@ static int cam_isp_packet_generic_blob_handler(void *user_data,
 
 		if (blob_size < (sizeof(uint32_t) * 2 + ubwc_config->num_ports *
 			sizeof(struct cam_ubwc_plane_cfg_v1) * 2)) {
-			CAM_ERR(CAM_ISP, "Invalid blob_size %u expected %u",
-				blob_size,
+			CAM_ERR(CAM_ISP, "Invalid blob_size %zu expected %zu",
+				(size_t)blob_size,
 				sizeof(uint32_t) * 2 + ubwc_config->num_ports *
 				sizeof(struct cam_ubwc_plane_cfg_v1) * 2);
 			return -EINVAL;
@@ -4181,8 +4181,8 @@ static int cam_isp_packet_generic_blob_handler(void *user_data,
 		struct cam_isp_csid_clock_config    *clock_config;
 
 		if (blob_size < sizeof(struct cam_isp_csid_clock_config)) {
-			CAM_ERR(CAM_ISP, "Invalid blob size %u expected %u",
-				blob_size,
+			CAM_ERR(CAM_ISP, "Invalid blob size %zu expected %zu",
+				(size_t)blob_size,
 				sizeof(struct cam_isp_csid_clock_config));
 			return -EINVAL;
 		}
@@ -4199,8 +4199,8 @@ static int cam_isp_packet_generic_blob_handler(void *user_data,
 		struct cam_fe_config *fe_config;
 
 		if (blob_size < sizeof(struct cam_fe_config)) {
-			CAM_ERR(CAM_ISP, "Invalid blob size %u expected %u",
-				blob_size, sizeof(struct cam_fe_config));
+			CAM_ERR(CAM_ISP, "Invalid blob size %zu expected %zu",
+				(size_t)blob_size, sizeof(struct cam_fe_config));
 			return -EINVAL;
 		}
 
@@ -4217,8 +4217,8 @@ static int cam_isp_packet_generic_blob_handler(void *user_data,
 			(struct cam_isp_init_frame_drop_config *)blob_data;
 
 		if (blob_size < sizeof(struct cam_isp_init_frame_drop_config)) {
-			CAM_ERR(CAM_ISP, "Invalid blob size %u expected %u",
-				blob_size,
+			CAM_ERR(CAM_ISP, "Invalid blob size %zu expected %zu",
+				(size_t)blob_size,
 				sizeof(struct cam_isp_init_frame_drop_config));
 			return -EINVAL;
 		}
@@ -4232,8 +4232,8 @@ static int cam_isp_packet_generic_blob_handler(void *user_data,
 		struct cam_isp_sensor_config *csid_dim_config;
 
 		if (blob_size < sizeof(struct cam_isp_sensor_config)) {
-			CAM_ERR(CAM_ISP, "Invalid blob size %u expected %u",
-				blob_size,
+			CAM_ERR(CAM_ISP, "Invalid blob size %zu expected %zu",
+				(size_t)blob_size,
 				sizeof(struct cam_isp_sensor_config));
 			return -EINVAL;
 		}
@@ -4253,8 +4253,8 @@ static int cam_isp_packet_generic_blob_handler(void *user_data,
 		struct cam_isp_prepare_hw_update_data   *prepare_hw_data;
 
 		if (blob_size < sizeof(struct cam_fps_config)) {
-			CAM_ERR(CAM_ISP, "Invalid fps blob size %u expected %u",
-				blob_size, sizeof(struct cam_fps_config));
+			CAM_ERR(CAM_ISP, "Invalid fps blob size %zu expected %zu",
+				(size_t)blob_size, sizeof(struct cam_fps_config));
 			return -EINVAL;
 		}
